requirements.txt
Flask~=3.0.3
pillow~=10.3.0
requests